Move components to scenes-app
