+++ packages/scenes/src/core/sceneGraph.ts
@@ -4,9 +4,8 @@ import { DefaultTimeRange, EmptyDataNode, EmptyVariableSet } from '../variables/
 import { sceneInterpolator } from '../variables/interpolation/sceneInterpolator';
 import { VariableCustomFormatterFn, SceneVariables } from '../variables/types';
 
-import { SceneDataProvider, SceneDataState, SceneLayout, SceneObject, SceneTimeRangeLike } from './types';
+import { SceneDataState, SceneLayout, SceneObject, SceneTimeRangeLike } from './types';
 import { lookupVariable } from '../variables/lookupVariable';
-import { SceneQueryRunner } from '../querying/SceneQueryRunner';
 
 /** Walks up the scene graph, returning the first non-undefined result of `extract` */
 function getClosest<T>(sceneObject: SceneObject, extract: (s: SceneObject) => T | undefined): T | undefined {
@@ -59,28 +58,6 @@ export function getLayout(scene: SceneObject): SceneLayout {
 }
 
 /**
- * Will walk up the scene object graph to the closest SceneQueryRunner
- */
-export function getSceneQueryRunner(sceneObject: SceneObject): SceneQueryRunner | undefined {
-  if (sceneObject.state.$data !== undefined) {
-    let curData: SceneDataProvider | undefined = sceneObject.state.$data;
-    while (curData !== undefined) {
-      if (curData instanceof SceneQueryRunner) {
-        return curData;
-      }
-
-      curData = curData.state.$data;
-    }
-  }
-  
-  if (sceneObject.parent) {
-    return getSceneQueryRunner(sceneObject.parent);
-  }
-
-  return undefined;
-}
-
-/**
  * Interpolates the given string using the current scene object as context.   *
  */
 export function interpolate(
@@ -122,7 +99,6 @@ export const sceneGraph = {
   getData,
   getTimeRange,
   getLayout,
-  getSceneQueryRunner,
   interpolate,
   lookupVariable,
   hasVariableDependencyInLoadingState,
